Broke up Archetypes into their own distinct classes - Finished various archetypes / noted various archetype 'bugs', added tertiary spellcaster slots
