+++ src/Classes/Paladin.ts
@@ -1,9 +1,7 @@
 import { PlayerClass, LevelingParams } from "./PlayerClass";
 import { PlayerCharacter } from "../Base/PlayerCharacter";
-import { ResourceTrait } from "../Base/Interfaces";
-import * as SpellList from "../../Assets/SpellList.json";
 import * as SpellcastingAbility from "../../Assets/SpellcastingAbility.json";
-import { PaladinArchetype } from "./Archetypes";
+import { PaladinArchetype } from "./Archetypes/PaladinArchetype";
 import { SpellSlotFactory } from "./SpellSlotFactory";
 
 export class Paladin extends PlayerClass {
